Manage secret vote
