+++ src/Products/PloneMeeting/browser/templates/itempolltype.pt
@@ -2,9 +2,11 @@
     <tal:loop repeat="pollType selectablePollTypes">
         <span class="fa fa-angle-double-right" />
         <a href="#"
-           tal:attributes="href python: '%s/@@change-item-polltype?new_value=%s'
-                           % (context.absolute_url(), pollType.token);
-                           class python: 'link-overlay-pm item_polltype_%s' % pollType.token"
+           tal:attributes="class python: 'link-overlay-pm item_polltype_%s' % pollType.token;
+                           onclick string:event.preventDefault();;
+                           callViewAndReload(base_url='${context/absolute_url}',
+                                             view_name='@@change-item-polltype',
+                                             params={'new_value': '${pollType/token}'});;;"
            tal:content="pollType/title">Poll type</a><br />
     </tal:loop>
 </tal:defines>
