+++ src/Products/PloneMeeting/validators.py
@@ -101,6 +101,7 @@ class DXCertifiedSignaturesValidator(validator.SimpleFieldValidator):
     def validate(self, value):
         """
         """
+        import ipdb; ipdb.set_trace()
         error = _validate_certified_signatures(value)
         if error:
             raise Invalid(error)
