+++ src/Products/PloneMeeting/widgets/pm_selectreadonly_input.pt
@@ -1,0 +1,35 @@
+<div xmlns="http://www.w3.org/1999/xhtml"
+     xmlns:tal="http://xml.zope.org/namespaces/tal"
+     tal:omit-tag="">
+<span id="" class=""
+      tal:attributes="id view/id;
+                      class view/klass;
+                      style view/style;
+                      title view/title;
+                      lang view/lang;
+                      onclick view/onclick;
+                      ondblclick view/ondblclick;
+                      onmousedown view/onmousedown;
+                      onmouseup view/onmouseup;
+                      onmouseover view/onmouseover;
+                      onmousemove view/onmousemove;
+                      onmouseout view/onmouseout;
+                      onkeypress view/onkeypress;
+                      onkeydown view/onkeydown;
+                      onkeyup view/onkeyup"><tal:block
+    repeat="value view/displayValue"
+    ><span class="selected-option"
+           tal:content="value"
+    /><tal:block condition="not:repeat/value/end">, </tal:block
+  ></tal:block
+></span>
+
+<tal:block define="id view/id;
+                   name string:${view/name}:list"
+           repeat="item view/items">
+  <input id="" name=""  value="" class="hidden-widget" type="hidden"
+         tal:condition="item/selected"
+         tal:attributes="id item/id; name name; value item/value" />
+</tal:block>
+
+</div>
